working on filtering WIP
