+++ src/lib/stores/filtered_movement_types.svelte.ts
@@ -1,0 +1,13 @@
+import { writable } from "svelte/store";
+
+export function createFilteredMovementTypesStore() {
+	const { subscribe, set, update } = writable([]);
+
+	return {
+		subscribe,
+		set,
+		update
+	};
+}
+
+export const filteredMovementTypes = createFilteredMovementTypesStore();
